catch Arrow exception and move tests
