+++ libs/libapi/tests/data/test_image_rgb.jpg
