+++ libs/libcommon/tests/conftest.py
@@ -5,12 +5,14 @@ from collections.abc import Iterator
 from environs import Env
 from pytest import fixture
 
+from libcommon.config import ParquetMetadataConfig
 from libcommon.queue import _clean_queue_database
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import _clean_cache_database
+from libcommon.storage import StrPath, init_parquet_metadata_dir
 
 # Import fixture modules as plugins
-pytest_plugins = ["tests.fixtures.datasets"]
+pytest_plugins = ["tests.fixtures.datasets", "tests.fixtures.fsspec"]
 
 
 @fixture(scope="session")
@@ -75,3 +77,8 @@ def cache_mongo_resource(cache_mongo_host: str) -> Iterator[CacheMongoResource]:
         yield cache_mongo_resource
         _clean_cache_database()
         cache_mongo_resource.release()
+
+
+@fixture
+def parquet_metadata_directory() -> StrPath:
+    return init_parquet_metadata_dir(ParquetMetadataConfig().storage_directory)
